Conflicts:
	intern/cycles/blender/blender_session.cpp
	source/blender/blenkernel/BKE_image.h
	source/blender/blenkernel/intern/image.c
	source/blender/blenkernel/intern/ocean.c
	source/blender/blenloader/intern/versioning_270.c
	source/blender/compositor/intern/COM_CompositorContext.h
	source/blender/compositor/intern/COM_Node.h
	source/blender/compositor/nodes/COM_CompositorNode.cpp
	source/blender/compositor/nodes/COM_ImageNode.cpp
	source/blender/compositor/nodes/COM_ImageNode.h
	source/blender/compositor/nodes/COM_OutputFileNode.cpp
	source/blender/compositor/nodes/COM_RenderLayersNode.cpp
	source/blender/compositor/operations/COM_OutputFileOperation.cpp
	source/blender/editors/render/render_opengl.c
	source/blender/editors/screen/screen_edit.c
	source/blender/editors/screen/screendump.c
	source/blender/editors/space_buttons/buttons_context.c
	source/blender/editors/space_image/image_ops.c
	source/blender/editors/space_node/node_edit.c
	source/blender/editors/space_view3d/view3d_draw.c
	source/blender/makesdna/DNA_scene_types.h
	source/blender/makesrna/intern/rna_scene_api.c
	source/blender/nodes/composite/nodes/node_composite_outputFile.c
	source/blender/render/extern/include/RE_pipeline.h
	source/blender/render/intern/source/pipeline.c
	source/blender/render/intern/source/render_result.c
	source/blender/render/intern/source/rendercore.c
	source/blender/windowmanager/wm.h

+++ source/blender/modifiers/intern/MOD_warp.c
@@ -34,7 +34,6 @@
 
 #include "BLI_math.h"
 #include "BLI_utildefines.h"
-#include "BLI_string.h"
 
 #include "BKE_cdderivedmesh.h"
 #include "BKE_modifier.h"
@@ -253,7 +252,7 @@ static void warpModifier_do(WarpModifierData *wmd, Object *ob,
 					fac = 3.0f * fac * fac - 2.0f * fac * fac * fac;
 					break;
 				case eWarp_Falloff_Root:
-					fac = (float)sqrt(fac);
+					fac = sqrtf(fac);
 					break;
 				case eWarp_Falloff_Linear:
 					/* pass */
@@ -262,7 +261,7 @@ static void warpModifier_do(WarpModifierData *wmd, Object *ob,
 					fac = 1.0f;
 					break;
 				case eWarp_Falloff_Sphere:
-					fac = (float)sqrt(2 * fac - fac * fac);
+					fac = sqrtf(2 * fac - fac * fac);
 					break;
 			}
